doc: add deprecation notice on timer lib cleanup
Due to an upcoming fix to allow the timer library to safely free its
allocations during the finalize() call[1], an ABI change will be
required. A new lock will be added to the rte_mem_config structure,
which will be used by the timer library to synchronize init/finalize
calls among multiple processes.

[1] http://patches.dpdk.org/patch/53334/

Signed-off-by: Erik Gabriel Carrillo <erik.g.carrillo@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>