app/testpmd: remove port stop check for macsec
There is no need to do such a check when set macsec for ixgbe,
reconfig_device_queueand is also useless. If we do not delete this
unnessary code, users have to sotp port before enable or disable
macsec, then restart this port after make configuration. All these
process is useless. As this cmdline is a private API which is only
used by ixgbe NIC, so remove it.

Fixes: 597f9fafe13b ("app/testpmd: convert to new Tx offloads API")
Cc: stable@dpdk.org

Signed-off-by: Wei Zhao <wei.zhao1@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>