net/i40e: add runtime option to disable vector rx
Vector RX functions are not at feature parity with non-vector ones and
currently the vector RX path is enabled by default. Hence, the only
option to force selection of non-vector variants and be able to retain
functionality is to disable vector PMD globally at compile time via the
config file option.

This patch introduces a new runtime device config option named
disable-vec-rx to allow users to limit the driver to make a choice among
non-vector RX functions, on a per device basis. This runtime option
defaults to a value of zero, allowing vector RX functions to be
selected (current behavior). When disable-vec-rx=1 is specified for a
device, even if all the other requirements to select a vector RX
function are satisfied, the driver will still pick one out of the
appropriate non-vector RX functions.

Signed-off-by: Mesut Ali Ergin <mesut.a.ergin@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>