kvargs: trim spaces at the beginning and end of key and values
When arguments delimited with RTE_KVARGS_PAIRS_DELIM happen to have
spaces before and after the delimiter, the rte_kvargs_process() can not
match them to the intended key. For example, the first argument below
triggers the support-multi-driver handler as expected, but the second
one silently fails to do so.

-w '84:00.0,support-multi-driver=1'
-w '84:00.0, support-multi-driver=1'

Signed-off-by: Mesut Ali Ergin <mesut.a.ergin@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>