meson: make build configurable
The first thing many developers do before start building DPDK is
disabling all the not needed divers and libraries. This happens
just because more than a half of DPDK dirvers and libraries are not
needed for the particular reason. For example, you don't need
dpaa*, octeon*, various croypto devices, eventdev, etc. if you're
only want to build OVS for x86_64 with static linking.

By disabling everything you don't need, build speeds up literally 10x
times. This is important for CI systems. For example, TravisCI wastes
10 minutes for the default DPDK build just to check linking with OVS.

Another thing is the binary size. Number of DPDK libraries and,
as a result, size of resulted statically linked application decreases
significantly.

Important thing also that you're able to not install some dependencies
if you don't have them on a target platform. Just disable libs/drivers
that depends on it. Similar thing for the glibc version mismatch
between build and target platforms.

Also, I have to note that less code means less probability of
failures and less number of attack vectors.

This patch gives 'meson' the power of configurability that we
have with 'make'. Using new options it's possible to enable just
what you need and nothing more.

For example, following cmdline could be used to build almost minimal
set of DPDK libs and drivers to check OVS build:

  $ meson build -Dexamples='' -Dtests=false -Denable_kmods=false \
                -Ddrivers_bus=pci,vdev          \
                -Ddrivers_mempool=ring          \
                -Ddrivers_net=null,virtio,ring  \
                -Ddrivers_crypto=virtio         \
                -Ddrivers_compress=none         \
                -Ddrivers_event=none            \
                -Ddrivers_baseband=none         \
                -Ddrivers_raw=none              \
                -Ddrivers_common=none           \
                -Dlibs=kvargs,eal,cmdline,ring,mempool,mbuf,net,meter,\
                       ethdev,pci,hash,cryptodev,pdump,vhost \
                -Dapps=none

Adding a few real net drivers will give configuration that can be used
in production environment.

Looks not very pretty, but this could be moved to a script.

Build details:

  Build targets in project: 57

  $ time ninja
  real    0m11,528s
  user    1m4,137s
  sys     0m4,935s

  $ du -sh ../dpdk_meson_install/
  3,5M    ../dpdk_meson_install/

To compare with what we have without these options:

  $ meson build -Dexamples='' -Dtests=false -Denable_kmods=false
  Build targets in project: 434

  $ time ninja
  real    1m38,963s
  user    10m18,624s
  sys     0m45,478s

  $ du -sh ../dpdk_meson_install/
  27M     ../dpdk_meson_install/

10x speed up for the user time.
7.7 times size decrease.

This is probably not much user-friendly because it's not a Kconfig
and dependency tracking in meson is really poor, so it requires
usually few iterations to pick correct set of libraries to satisfy
all dependencies. However, it's not a big deal. Options intended
for a proficient users who knows what they need.

Signed-off-by: Ilya Maximets <i.maximets@samsung.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>