eal: remove non-thread panic calls from init sequence
This patch changes some void functions to return a value,
so that the init sequence may tear down orderly
instead of calling panic.
The calls for launching core messaging threads were left in tact
in all 3 eal implementations.
This should be addressed in a different patchset.
There are still cases where the PMDs or message threads
may call panic with no context for the user.
For these I will submit a patch suggesting a callback registration,
allowing the user to register a context to be called
in case of a panic that takes place outside the init sequence.

Signed-off-by: Arnon Warshavsky <arnon@qwilt.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>