net/virtio: fix segmented packet issue in mergeable Rx path
After having dequeued a burst of descriptors, there may be a
need to dequeue a few more if the last packet was segmented
and not complete.

When it happens, the extra segments were not properly attached
to the mbuf chain, and so were lost.

This patch fixes the mbuf chaining.

Fixes: bcac5aa207f8 ("net/virtio: improve batching in mergeable path")
Cc: stable@dpdk.org

Reported-by: Yaroslav Brustinov <ybrustin@cisco.com>
Signed-off-by: Maxime Coquelin <maxime.coquelin@redhat.com>

Signed-off-by: Maxime Coquelin <maxime.coquelin@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>