net/ixgbevf: add an option pflink_fullchk to get link status quickly
In some scenario, if the PF's running status is stopped, the VF link
status can be detected as down by checking the mailbox status. This
fully checking will introduce mailbox interrupt in PF, and it will be
bad if many VFs are running and the application is checking the VF's
link status quickly (wait_to_complete is 0) and frequently.

Normally, checking the PF's link status is enough, no need to fully
check. For different scenarios, add 'pflink_fullchk' option to control
whether check the link fully or not.

Fixes: 91546fb62e67 ("net/ixgbevf: fix link state")
Cc: stable@dpdk.org

Signed-off-by: Haiyue Wang <haiyue.wang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>