crypto/qat: handle buffer size for digest-encrypted auth-cipher
This patch adds condition to be met when using
wireless algorithms (SNOW3G, KASUMI, ZUC) in
out-of-place auth-cipher operations. It checks
if the digest location overlaps with the data to
be encrypted or decrypted and if so, treats as a
digest-encrypted case.
It also checks if the digest is being encrypted
or decrypted partially and extends PMD buffers
accordingly.

Signed-off-by: Damian Nowak <damianx.nowak@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>