net/af_xdp: support need wakeup feature
This patch adds a new devarg to support the need_wakeup flag for Tx and
fill rings, when this flag is set by the driver, it means that the
userspace application has to explicitly wake up the kernel Rx or kernel Tx
processing by issuing a syscall. Poll() can wake up both and sendto() or
its alternatives will wake up Tx processing only.

This feature is to provide efficient support for case that application and
driver are executing on the same core.

Signed-off-by: Xiaolong Ye <xiaolong.ye@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>