examples/multi_process - fix crash in mp_client with sparse ports
The mp_client crashes if run on Azure or any system where ethdev
ports are owned. This is because the flush loop is confusing the
index into the array of ports[] with the port id.

For example if the server has Ports 3 and 5. Then calling
rte_eth_tx_buffer_flush on any other buffer will dereference null
because the tx buffer for that port was not allocated.

v2
   - fix typo
   - the flush code is common enough that it should not be marked unlikely
   - combine conditions to reduce indentation

v3
   - avoid unnecessary if() if sent is zero.

Fixes: e2366e74e029 ("examples: use buffered Tx")
Signed-off-by: Stephen Hemminger <sthemmin@microsoft.com>
Signed-off-by: Stephen Hemminger <sthemmin@microsoft.com>
Reviewed-by: David Marchand <david.marchand@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>