eal: rename state values in rte_lcore_state
C language does not really treat enum's as first class symbols.
The values in an enum live in a global namespace.  That means if
DPDK defines "RUNNING" it can't be used by another enum in an
application using DPDK.

To solve this add a prefix "RTE_LCORE_" to the enum values, and
make them grammatically consistent.

Also, simplify the inline thread_is_running() which is copied
in softnic and ip_pipeline.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>