net/e1000: i219 unit hang issue fix on reset/close
Unit hang may occur if multiple descriptors are available in the rings
during reset or close. This state can be detected by configure status
by bit 8 in register. If the bit is set and there are pending descriptors
in one of the rings, we must flush them before reset or close.

Signed-off-by: Xiao Zhang <xiao.zhang@intel.com>
Acked-by: Wei Zhao <wei.zhao1@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>