examples/multi_process/client_server_mp: check port validity
The mp_server incorrectly allows a port mask that included hidden
ports and which later caused either lost packets or failed initialization.

This fixes explicitly checking that each bit in portmask is a
valid port before using it.

The max_ports parameter is no longer necessary, so remove it
from the caller.

Fixes: 5b7ba31148a8 ("ethdev: add port ownership")
Signed-off-by: Stephen Hemminger <sthemmin@microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>