examples/power: fix policy handling for fifo
While the core frequency scale-up and scale down all works fine with the
fifo-per-core functionaliry, there was a gap for policy handling. When
creating or destroying a policy, the core associated with the fifo
needs to be automatically added to the core list, so the policy is
associated with the correct core.

Fixes: 221e7026d521 ("examples/power: add FIFO per core for JSON interface")

Signed-off-by: David Hunt <david.hunt@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>