net/i40e: fix SFP X722 not work with FW4.16
The code checked the NVM API version of FW to check if support adminq
operation, when version is 1.7 or above adminq operation is support
and the adminq flag will be set. The code will use adminq to set TPID if
the flags set or to use register. The NVM API version of SFP X722 old
versions are lowwer than 1.7 and for FW4.16 is 1.8. But FW4.16 of SFP
X722 does not support adminq operation, the legacy logic is not compatible
with device SFP X722 currently. Check the device id to unset adminq
flag to use resigter setting to make it work.

Fixes: 73cd7d6dc8e1 ("net/i40e: use set switch AQ instead of register
setting")
Cc: stable@dpdk.org

Signed-off-by: Xiao Zhang <xiao.zhang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>