net/ixgbe: fix x550 Reta set fail error
There is a bug in function ixgbevf_dev_info_get(), it do not
contain the reta table size get function for vf port, and this
will cause error when update reta for x550 vf port. And also
when function ixgbe_reta_size_get() is called, it should return
other ixgbe NIC vf reta support or not, not only x550.

Cc: stable@dpdk.org

Fixes: 2144f6630fca ("ixgbe: add redirection table size in device info")
Signed-off-by: wei zhao <wei.zhao1@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>